Gave better error for CAPTCHA
I was having trouble logging into jira, and it took me a long time to
troubleshoot the issue. It turned out that my password manager was not
pulling authentication details properly. After this, I was locked out
because of the CAPTCHA. I was not seeing the original error message,
which would have been very helpful.

Changing the `echo` to an `echoe` made the message appear properly.