Only use 'open' or 'xdg-open' if they are executable
When using :ViraBrowse to open the current issue in the default browser
Vira will only try to use 'open' or 'xdg-open' if they exist and are
executable. This is a fallback for case where users have not provided
'g:vira_browser' or the 'BROWSER' environment variable.

Given a config/system where 'g:vira_browser' and 'BROWSER' are not set
and 'open', 'xdg-open' are not executable, an informative error will be
displayed asking the user to provide one of the two overriding
variables.

When provided with a command via '$BROWSER' or 'g:vira_browser' that is
not executable, :ViraBrowse will now display an informative error to the
user.