Don't skip project config if defaults are missing
If projects are defined according to `README.md` in
`vira_projects.json`, then the project configuration was still not
loaded because the implementation assumed that `__default__` must exist.

This commit changes the `load_project_config()` such that it cares about
`__default__` section only if the project name could not be determined
based on the input parameter nor the Git repository.