VIRA-276: Handle missing assignee and guess current user correctly
In few places the implementation assumed that the assignee field is
present in every Jira issue returned which is not necessarily true and
in fact seen otherwise in a fairly big production Jira instance.  This
is fixed uniformly by checking field's presence.

The current user was also ambiguously determined which in some cases
returned incorrect result:  if the 'assignee or reporter' query returned
the current user as the reporter first, the implementation still checked
the assignee first and assumed it to be the current user (which could be
different that the reporter being the current user).  The fix is to
query the assignee and reporter _separately_ to determine the current
user.