Don't append Null in non-empty server list
By default, 'Null' is appended when a list of servers is displayed.
This makes sense if there are no servers configured to not leave the
list blank, but looks somewhat confusing if there are servers available.

This commit puts 'Null' in the server list only if no servers are
available (configured).